rename app packages
