+++ app/src/main/java/me/yifeiyuan/flapdev/testcases/HeaderFooterTestcase.kt
@@ -1,6 +1,5 @@
-package me.yifeiyuan.flapdev.showcase
+package me.yifeiyuan.flapdev.testcases
 
-import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
