+++ app/src/main/java/me/yifeiyuan/flapdev/testcases/ItemDecorationTestcase.kt
@@ -1,0 +1,17 @@
+package me.yifeiyuan.flapdev.testcases
+
+import android.view.View
+
+/**
+ * Created by 程序亦非猿 on 2022/8/18.
+ */
+
+class ItemDecorationTestcase : BaseTestcaseFragment() {
+
+
+    override fun onInit(view: View) {
+        super.onInit(view)
+
+    }
+
+}
