+++ app/src/main/java/me/yifeiyuan/flapdev/testcases/LayoutDelegateDSLTestcase.kt
@@ -1,4 +1,4 @@
-package me.yifeiyuan.flapdev.showcase
+package me.yifeiyuan.flapdev.testcases
 
 import android.util.Log
 import android.view.View
