+++ app/src/main/java/me/yifeiyuan/flapdev/testcases/MultiTypeTestcase.kt
@@ -1,4 +1,4 @@
-package me.yifeiyuan.flapdev.showcase
+package me.yifeiyuan.flapdev.testcases
 
 import android.view.Menu
 import android.view.MenuInflater
