+++ app/src/main/java/me/yifeiyuan/flapdev/testcases/ViewPager2Testcase.kt
@@ -1,4 +1,4 @@
-package me.yifeiyuan.flapdev.showcase
+package me.yifeiyuan.flapdev.testcases
 
 import android.os.Bundle
 import android.view.*
