+++ app/src/main/res/layout/fragment_view_pager2.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".showcase.ViewPager2Testcase">
+    tools:context=".testcases.ViewPager2Testcase">
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/viewPager"
