+++ flap/src/main/java/me/yifeiyuan/flap/ext/ExtraParamsProvider.kt
@@ -8,3 +8,12 @@ package me.yifeiyuan.flap.ext
 interface ExtraParamsProvider {
     fun getParam(key: String): Any?
 }
+
+/**
+ * @since 3.0.4
+ */
+internal class ExtraParamsProviderWrapper(private val block: (key: String) -> Any?) : ExtraParamsProvider {
+    override fun getParam(key: String): Any? {
+        return block(key)
+    }
+}
